create simple product and composite product
