+++ packages/extension-chakra-storefront/src/pages/product-detail/partials/simple-product-details.jsx
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2024, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import React, {Fragment} from 'react'
+import PropTypes from 'prop-types'
+import ProductView from '../../../components/product-view'
+import InformationAccordion from './information-accordion'
+
+const SimpleProductDetails = ({
+    product,
+    primaryCategory,
+    isProductLoading,
+    isBasketLoading,
+    isWishlistLoading,
+    handleAddToWishlist,
+    handleAddToCart
+}) => {
+    return (
+        <Fragment>
+            <ProductView
+                product={product}
+                category={primaryCategory?.parentCategoryTree || []}
+                addToCart={(variant, quantity) => handleAddToCart([{product, variant, quantity}])}
+                addToWishlist={handleAddToWishlist}
+                isProductLoading={isProductLoading}
+                isBasketLoading={isBasketLoading}
+                isWishlistLoading={isWishlistLoading}
+            />
+            <InformationAccordion product={product} />
+        </Fragment>
+    )
+}
+
+SimpleProductDetails.propTypes = {
+    product: PropTypes.object,
+    primaryCategory: PropTypes.object,
+    isProductLoading: PropTypes.bool,
+    isBasketLoading: PropTypes.bool,
+    isWishlistLoading: PropTypes.bool,
+    handleAddToWishlist: PropTypes.func,
+    handleAddToCart: PropTypes.func
+}
+
+export default SimpleProductDetails
